Updated up command for reuse
